Split up into small test files for each exercise
