+++ package.json
@@ -4,7 +4,7 @@
   "description": "Exercises to learn and practice the `Array` methods:",
   "main": "index.js",
   "scripts": {
-    "test": "NODE_OPTIONS=--experimental-vm-modules npx jest --watchAll"
+    "test": "NODE_OPTIONS=--experimental-vm-modules npx jest --watch -o"
   },
   "repository": {
     "type": "git",
